Resolve rebase issues, restructure code and fix warnings
