--- swap/src/io.rs
@@ -1,8 +1,0 @@
-// This struct contains all the I/O required to execute a swap
-pub struct Io {
-    // swarm: libp2p::Swarm<>,
-// bitcoind_rpc: _,
-// monerod_rpc: _,
-// monero_wallet_rpc: _,
-// db: _,
-}
