+++ swap/src/lib.rs
@@ -6,9 +6,7 @@ use std::fmt::{self, Display};
 pub mod alice;
 pub mod bitcoin;
 pub mod bob;
-pub mod bob_simple;
 pub mod cli;
-pub mod io;
 pub mod monero;
 pub mod network;
 pub mod recover;
