+++ xmr-btc/src/bob.rs
@@ -879,63 +879,3 @@ impl State5 {
         self.tx_lock.txid()
     }
 }
-
-/// Watch for the refund transaction on the blockchain. Watch until t2 has
-/// elapsed.
-pub async fn watch_for_refund_btc<W>(state: State5, bitcoin_wallet: &W) -> Result<()>
-where
-    W: WatchForRawTransaction,
-{
-    let tx_cancel = bitcoin::TxCancel::new(
-        &state.tx_lock,
-        state.refund_timelock,
-        state.A.clone(),
-        state.b.public(),
-    );
-
-    let tx_refund = bitcoin::TxRefund::new(&tx_cancel, &state.refund_address);
-
-    let tx_refund_watcher = bitcoin_wallet.watch_for_raw_transaction(tx_refund.txid());
-
-    Ok(())
-}
-
-// Watch for refund transaction on the blockchain
-pub async fn watch_for_redeem_btc<W>(state: State4, bitcoin_wallet: &W) -> Result<State5>
-where
-    W: WatchForRawTransaction,
-{
-    let tx_redeem = bitcoin::TxRedeem::new(&state.tx_lock, &state.redeem_address);
-    let tx_redeem_encsig = state
-        .b
-        .encsign(state.S_a_bitcoin.clone(), tx_redeem.digest());
-
-    let tx_redeem_candidate = bitcoin_wallet
-        .watch_for_raw_transaction(tx_redeem.txid())
-        .await;
-
-    let tx_redeem_sig =
-        tx_redeem.extract_signature_by_key(tx_redeem_candidate, state.b.public())?;
-    let s_a = bitcoin::recover(state.S_a_bitcoin.clone(), tx_redeem_sig, tx_redeem_encsig)?;
-    let s_a = monero::private_key_from_secp256k1_scalar(s_a.into());
-
-    Ok(State5 {
-        A: state.A.clone(),
-        b: state.b.clone(),
-        s_a,
-        s_b: state.s_b,
-        S_a_monero: state.S_a_monero,
-        S_a_bitcoin: state.S_a_bitcoin.clone(),
-        v: state.v,
-        btc: state.btc,
-        xmr: state.xmr,
-        refund_timelock: state.refund_timelock,
-        punish_timelock: state.punish_timelock,
-        refund_address: state.refund_address.clone(),
-        redeem_address: state.redeem_address.clone(),
-        punish_address: state.punish_address.clone(),
-        tx_lock: state.tx_lock.clone(),
-        tx_refund_encsig: state.tx_refund_encsig.clone(),
-        tx_cancel_sig: state.tx_cancel_sig_a.clone(),
-    })
-}
